% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state-panel.R
\name{attempt_date}
\alias{attempt_date}
\title{Attempt to convert input to date}
\usage{
attempt_date(date, by)
}
\arguments{
\item{date}{Date}

\item{by}{Temporal resolution}
}
\description{
Attempt to convert input to date
}
