% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{country_names}
\alias{country_names}
\title{Country names}
\usage{
country_names(x, list = "GW")
}
\arguments{
\item{x}{A vector of numeric country codes}

\item{list}{Which states list to use? Only "GW" at this time.}
}
\description{
Country names
}
\examples{
data("gwstates")
codes <- gwstates$gwcode
cn    <- country_names(codes)
data.frame(gwcode = codes, country_name = cn)

}
