% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-missing.R
\name{.warner}
\alias{.warner}
\title{Temporary helper function to warn about argument order change}
\usage{
.warner(data, x)
}
\arguments{
\item{data}{data.frame}

\item{x}{string}
}
\description{
Temporary helper function to warn about argument order change
}
\keyword{internal}
