% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index-date.R
\name{index_date}
\alias{index_date}
\title{Index date}
\usage{
index_date(x, period)
}
\arguments{
\item{x}{a \code{\link[base:Dates]{base::Date()}} vector}

\item{period}{a time period, e.g. "year" or "month"}
}
\description{
Find the first day of the time intervals in which a set of dates fall
}
\keyword{internal}
