% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-missing.R
\name{mssng_mat_base}
\alias{mssng_mat_base}
\title{Compute matrix of missing status for state panel}
\usage{
mssng_mat_base(x, data, space, time, time_unit, statelist = "none")
}
\arguments{
\item{x}{Variable names(s), e.g. "x" or c("x1", "x2").}

\item{data}{State panel data frame}

\item{space}{Name of state identifier}

\item{time}{Name of time identifier}

\item{time_unit}{Temporal resolution character string, e.g. "year" or "month".
See details in \code{\link[base]{seq.Date}}.}

\item{statelist}{Check not only missing values, but presence or absence of
observations against a list of independent states?}
}
\description{
Compute matrix of missing status for state panel
}
