% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_date}
\alias{parse_date}
\title{Parse date}
\usage{
parse_date(x)
}
\arguments{
\item{x}{a integer or character vector, see examples}
}
\description{
Try to parse input as a date
}
\examples{
parse_date(2006)
parse_date("2006")
parse_date("2006-06")
parse_date("2006-06-01")
parse_date(as.Date("2006-06-01"))

}
