% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-missing.R
\name{plot_missing}
\alias{plot_missing}
\title{Visualize missing status for state panel}
\usage{
plot_missing(x, data, space, time, time_unit, statelist = c("none", "GW",
  "COW"))
}
\arguments{
\item{x}{Variable names(s), e.g. "x" or c("x1", "x2"). To check all variables
in a data frame use something like.}

\item{data}{State panel data frame}

\item{space}{Name of state identifier}

\item{time}{Name of time identifier}

\item{time_unit}{Temporal resolution character string, e.g. "year" or "month".
See details in \code{\link[base]{seq.Date}}.}

\item{statelist}{Check not only missing values, but presence or absence of
observations against a list of independent states?}
}
\description{
Visualize missing status for state panel
}
\examples{
# To check all variables:
# plot_missing(setdiff(colnames(df), "space", "time"), ...)

}
