% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pretty_names}
\alias{pretty_names}
\title{Shorten country names}
\usage{
pretty_names(x)
}
\arguments{
\item{x}{([character])\cr
Country names, e.g. from [country_names()].}
}
\description{
Shorten country names
}
\examples{
cn <- c(
  "Macedonia, the former Yugoslav Republic of",
  "Congo, the Democratic Republic of the",
  "Tanzania, United Republic of")
pretty_names(cn)

}
