% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sfind}
\alias{sfind}
\title{Lookup country codes or names}
\usage{
sfind(x, list = "both")
}
\arguments{
\item{x}{The search string or number.}

\item{list}{Which state list to search. Only G&W at this point.}
}
\description{
Helper to look up state list entries by country code or name
}
\examples{
sfind(325)
sfind("Algeria")

}
