% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_panel.R
\name{state_panel_date}
\alias{state_panel_date}
\title{State panel constructor}
\usage{
state_panel_date(start, end, by, partial, useGW)
}
\arguments{
\item{start}{length 1 date}

\item{end}{length 1 date}

\item{by}{time period}

\item{partial}{how to handle partial interval overlap}

\item{useGW}{Use G&W statelist (\code{TRUE}), or COW (\code{FALSE})?}
}
\description{
Internal state panel constructor without input checking and fluff to allow
argument shortcuts like start = 2000 instead of a full date
}
\keyword{internal}
