% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states-package.R
\docType{package}
\name{states}
\alias{states}
\alias{states-package}
\title{State system membership}
\description{
Create data based on the Gleditsch & Ward (G&W) or Correlates of
War (COW) state system memberships lists. This is useful as a template
for merging other sources of data that have conflicting sets of states.
}
\details{
See static docs at https://andybeger.com/states and the source code
at https://www.github.com/andybega/states
}
\references{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Gleditsch, Kristian S. & Michael D. Ward. 1999. ``Interstate System
  Membership: A Revised List of the Independent States since 1816.''
  International Interactions 25: 393-413.

Correlates of War Project. 2017. ``State System Membership List, v2016.''
  Online, http://correlatesofwar.org
}\if{html}{\out{</div>}}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andybega/states}
  \item \url{https://www.andybeger.com/states/}
  \item Report bugs at \url{https://github.com/andybega/states/issues}
}

}
\author{
\strong{Maintainer}: Andreas Beger \email{adbeger@gmail.com} (\href{https://orcid.org/0000-0003-1883-3169}{ORCID})

}
