% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sfind}
\alias{sfind}
\title{Lookup country codes or names}
\usage{
sfind(x, list = "both")
}
\arguments{
\item{x}{The search string or number.}

\item{list}{Which state list to search (both, G&W, or COW only)}
}
\description{
Helper to look up state list entries by country code or name
}
\examples{
# Works with either integer or strings
sfind(325)
sfind("ALG")
sfind("Algeria")

# Search strings are treated as regular expressions (see stringr::str_detect)
sfind("Germany")
sfind("German")
}
