% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state-panel.R
\name{state_panel}
\alias{state_panel}
\title{Create state panel data}
\usage{
state_panel(start_date, end_date, by = "month", useGW = TRUE)
}
\arguments{
\item{start_date}{Beginning date for data}

\item{end_date}{End date for data}

\item{by}{Temporal resolution}

\item{useGW}{Use Gleditsch & Ward statelist or Correlates of War state system
membership list.}
}
\description{
Create panel data consisting of independent states in the international system.
}
\examples{
begin <- as.Date("1991-01-01")
end   <- as.Date("2015-01-01")
gwlist  <- state_panel(begin, end, by = "year", useGW = TRUE)
cowlist <- state_panel(begin, end, by = "year", useGW = FALSE)

}
